refactor(web): replace useOrganizationProjects with granular hooks for project management
- Updated project-related components to utilize new hooks: `useCreateOrganizationProject`, `useEditOrganizationProject`, `useDeleteOrganizationProject`, and `useOrganizationProjectsQuery`.
- Removed the deprecated `useOrganizationProjects` hook, enhancing clarity and maintainability in project management logic.
- Improved data fetching and state management by leveraging specialized hooks for creating, editing, and deleting organization projects.
