+++ apps/web/core/components/features/projects/bulk-actions/bulk-archive-projects-modal.tsx
@@ -1,4 +1,5 @@
-import { useOrganizationProjects } from '@/core/hooks';
+import { useEditOrganizationProject } from '@/core/hooks/organizations/projects/use-edit-organization-project';
+import { useOrganizationProjectsQuery } from '@/core/hooks/organizations/projects/use-organization-projects-query';
 import { useAtomValue } from 'jotai';
 import { Button, Modal, Text } from '@/core/components';
 import { useTranslations } from 'next-intl';
@@ -27,7 +28,8 @@ export function BulkArchiveProjectsModal(props: IBulkArchiveProjectModalProps) {
 	const { open, closeModal, projectIds = [] } = props;
 	const organizationProjects = useAtomValue(organizationProjectsState);
 
-	const { setOrganizationProjects, editOrganizationProject, getOrganizationProjects } = useOrganizationProjects();
+	const { setOrganizationProjects, getOrganizationProjects } = useOrganizationProjectsQuery();
+	const { editOrganizationProject } = useEditOrganizationProject();
 	const [isLoading, setIsLoading] = useState(false);
 
 	const projects = useMemo(
