+++ apps/web/core/hooks/organizations/index.ts
@@ -1,6 +1,12 @@
 export * from './employees/use-productivity-table-config';
 export * from './employees/use-employee';
 export * from './projects/use-organization-projects';
+export * from './projects/use-invalidate-organization-projects';
+export * from './projects/use-organization-projects-query';
+export * from './projects/use-create-organization-project';
+export * from './projects/use-edit-organization-project';
+export * from './projects/use-delete-organization-project';
+export * from './projects/use-organization-projects-pagination';
 export * from './teams/use-active-team';
 export * from './teams/use-auth-team-tasks';
 export * from './teams/use-organization-teams-employee';
